Update README.md
The evolving Takagi-Sugeno (eTS) is a forecasting model classified as a rule-based evolving Fuzzy System (eFS) proposed by Angelov and Filev [1].

The eTS model consists of two main parts, the antecedent, and the consequent. The antecedent part comprehends form rules from the input space. Each rule is a cluster of similar inputs. Each rule has associated with a set of parameters responsible for computing the model's output. The eTS model uses weighted Recursive Least Square (wRLS) to update the parameters of the consequent. The paper [2] has a complete review of the eFSs.

[1] P. Angelov, D. Filev, On-line design of takagi-sugeno models, in: International Fuzzy Systems Association World Congress, Springer, 2003, pp. 576–584, https://doi.org/10.1007/3-540-44967-1_69.

[2] K.S.T.R. Alves, E. P. Aguiar, A novel rule-based evolving Fuzzy System applied to the thermal modeling of power transformers. Appl. Soft Comput. 112 (2021) 107764-107764, https://doi.org/10.1016/j.asoc.2021.